feat: header 구조 변경
